#53 moving aes types in their own folder
