+++ cas-dotnet-sdk/Symmetric/Types/AesBytesDecrypt.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CasDotnetSdk.Symmetric.Types
+{
+    internal struct AesBytesDecrypt
+    {
+        public IntPtr plaintext { get; set; }
+        public int length { get; set; }
+    }
+}
