+++ cas-dotnet-sdk/Symmetric/Types/AesBytesEncrypt.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CasDotnetSdk.Symmetric.Types
+{
+    internal struct AesBytesEncrypt
+    {
+        public IntPtr ciphertext { get; set; }
+        public int length { get; set; }
+    }
+}
