+++ cas-dotnet-sdk/Symmetric/Types/AesEncryptResult.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CasDotnetSdk.Symmetric.Types
+{
+    public class AesEncryptResult
+    {
+        public string Key { get; set; }
+        public string CipherText { get; set; }
+    }
+}
